net/https: do not systematically use tlsCert as CAfile.
For now, on client/server interaction, if you want to use tls on client
side, you have to activate with tlsOn and tlsCert is mandatory.
This field will be used to set your CApath. It is this file that will
be used to validate server certificate.
Disable this fonctionality by default.

With this patch, if the user specifies tlsCert, the comportment will be
the same than previously. But, if it is not indicated, default CApaths
will be used.

Priv-Id: eb83da66456990ee65e5c3e67205f0c74fb830ba

+++ src/net/http.c
@@ -2538,14 +2538,7 @@ int httpc_cfg_from_iop(httpc_cfg_t *cfg, const core__httpc_cfg__t *iop_cfg)
 
     if (iop_cfg->tls_on) {
         SB_1k(err);
-        char path[PATH_MAX] = "/tmp/tls-cert-XXXXXX";
-        int ret;
         core__tls_cert_and_key__t *data;
-
-        if (!iop_cfg->tls_cert.s) {
-            logger_error(&_G.logger, "tls: no certificate provided");
-            return -1;
-        }
 
         if (iop_cfg->tls_client) {
             data = IOP_UNION_GET(core__tls_cfg, iop_cfg->tls_client, data);
@@ -2563,23 +2556,30 @@ int httpc_cfg_from_iop(httpc_cfg_t *cfg, const core__httpc_cfg__t *iop_cfg)
             return -1;
         }
 
-        ret = write_in_tmp_file(path, iop_cfg->tls_cert.s,
-                                iop_cfg->tls_cert.len, &err);
+        if (iop_cfg->tls_cert.s) {
+            char path[PATH_MAX] = "/tmp/tls-cert-XXXXXX";
+            int ret;
 
-        if (ret < 0) {
-            httpc_cfg_tls_wipe(cfg);
-            logger_error(&_G.logger, "tls: failed to dump certificate: "
-                         "%*pM", SB_FMT_ARG(&err));
-            return -1;
-        }
+            ret = write_in_tmp_file(path, iop_cfg->tls_cert.s,
+                                    iop_cfg->tls_cert.len, &err);
 
-        ret = httpc_cfg_tls_add_verify_file(cfg, LSTR(path));
-        unlink(path);
+            if (ret < 0) {
+                httpc_cfg_tls_wipe(cfg);
+                logger_error(&_G.logger, "tls: failed to dump certificate: "
+                             "%*pM", SB_FMT_ARG(&err));
+                return -1;
+            }
 
-        if (ret < 0) {
-            httpc_cfg_tls_wipe(cfg);
-            logger_error(&_G.logger, "tls: failed to load certificate");
-            return -1;
+            ret = httpc_cfg_tls_add_verify_file(cfg, LSTR(path));
+            unlink(path);
+
+            if (ret < 0) {
+                httpc_cfg_tls_wipe(cfg);
+                logger_error(&_G.logger, "tls: failed to load certificate");
+                return -1;
+            }
+        } else {
+            SSL_CTX_set_default_verify_paths(cfg->ssl_ctx);
         }
     }
